+++ lib/widgets/base/com_arrow.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chen_common/common/style.dart';
+import 'package:get/get.dart';
 
 class ComArrow extends StatelessWidget {
   final double? size;
@@ -14,7 +14,26 @@ class ComArrow extends StatelessWidget {
     return Icon(
       isBack ? Icons.arrow_back_ios : Icons.arrow_forward_ios,
       size: size ?? CommonStyle.fontMd,
-      color: color ?? CupertinoColors.systemGrey2.resolveFrom(context),
+      color: color ?? CommonColors.theme.shade300,
     );
+  }
+}
+
+class ComBack extends StatelessWidget {
+  final Color? color;
+
+  const ComBack({super.key, this.color});
+
+  @override
+  Widget build(BuildContext context) {
+    return !Navigator.of(context).canPop()
+        ? Container()
+        : GestureDetector(
+            child: Icon(
+              Icons.arrow_back_ios,
+              color: color ?? CommonColors.theme.shade400,
+            ),
+            onTap: () => Get.back(),
+          );
   }
 }